requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
langchain==0.2.16
langchain-core==0.2.43
langchain-groq==0.1.9
langsmith==0.1.147
groq==0.9.0
snowflake-connector-python==3.10.0
pyyaml==6.0.1
python-dotenv==1.0.1
pandas==2.2.2